Rename IExprVisitor2 and move LispExamplesTest
