+++ jproto/src/main/java/uk/bs338/hashLisp/jproto/eval/expr/ExprFactory.java
@@ -126,7 +126,7 @@ public class ExprFactory {
             return true;
         }
 
-        @Override public <V extends IExprVisitor2> @NotNull V visit(@NotNull V visitor) {
+        @Override public <V extends IExprVisitor> @NotNull V visit(@NotNull V visitor) {
             visitor.visitSimple(this);
             return visitor;
         }
@@ -142,7 +142,7 @@ public class ExprFactory {
             return true;
         }
 
-        @Override public <V extends IExprVisitor2> @NotNull V visit(@NotNull V visitor) {
+        @Override public <V extends IExprVisitor> @NotNull V visit(@NotNull V visitor) {
             visitor.visitSymbol(this);
             return visitor;
         }
@@ -194,7 +194,7 @@ public class ExprFactory {
             return true;
         }
 
-        @Override public <V extends IExprVisitor2> @NotNull V visit(@NotNull V visitor) {
+        @Override public <V extends IExprVisitor> @NotNull V visit(@NotNull V visitor) {
             visitor.visitCons(this);
             return visitor;
         }
